Add widget layout methods.
The new widget layout methods will constraint the widgets based on
widgets are positioned together. This will automatically add the margin
of the current theme.

Possibly in the future, this will allow widget layout to be flipped for
right-to-left language localization.

+++ GUI/include/TTauri/GUI/Window_base.hpp
@@ -403,8 +403,8 @@ private:
         widgetSolver.suggest(widget->box.height, 0);
         minimumWindowExtent = widget->box.currentExtent();
 
-        widgetSolver.suggest(widget->box.width, static_cast<double>(std::numeric_limits<uint32_t>::max()));
-        widgetSolver.suggest(widget->box.height, static_cast<double>(std::numeric_limits<uint32_t>::max()));
+        widgetSolver.suggest(widget->box.width, std::numeric_limits<uint32_t>::max());
+        widgetSolver.suggest(widget->box.height, std::numeric_limits<uint32_t>::max());
         maximumWindowExtent = widget->box.currentExtent();
 
         LOG_INFO("Window '{}' minimumExtent={} maximumExtent={}", title, minimumWindowExtent, maximumWindowExtent);
